+++ test/videostreamindex/DefaultVideoStreamIndex.test.ts
@@ -15,6 +15,7 @@ import {
   SdkTrackMapping,
 } from '../../src/signalingprotocol/SignalingProtocol.js';
 import DefaultVideoStreamIndex from '../../src/videostreamindex/DefaultVideoStreamIndex';
+import VideoSendingAttendee from '../../src/videostreamindex/VideoSendingAttendee';
 
 describe('DefaultVideoStreamIndex', () => {
   let expect: Chai.ExpectStatic;
@@ -221,10 +222,13 @@ describe('DefaultVideoStreamIndex', () => {
   });
 
   describe('allVideoSendingAttendeesExcludingSelf', () => {
+    const compare = (a: VideoSendingAttendee, b: VideoSendingAttendee): number =>
+      a.attendeeId.localeCompare(b.attendeeId);
+
     it('returns set of all video sending attendees', () => {
-      expect(Array.from(index.allVideoSendingAttendeesExcludingSelf('xy1'))).to.deep.equal([]);
+      expect(index.allVideoSendingAttendeesExcludingSelf('xy1')).to.deep.equal([]);
       index.integrateIndexFrame(new SdkIndexFrame());
-      expect(Array.from(index.allVideoSendingAttendeesExcludingSelf('xy1'))).to.deep.equal([]);
+      expect(index.allVideoSendingAttendeesExcludingSelf('xy1')).to.deep.equal([]);
       index.integrateIndexFrame(
         new SdkIndexFrame({
           sources: [
@@ -233,6 +237,7 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 1,
               maxBitrateKbps: 100,
               attendeeId: 'xy1',
+              externalUserId: 'ext#xy1',
               mediaType: SdkStreamMediaType.VIDEO,
             }),
             new SdkStreamDescriptor({
@@ -240,6 +245,7 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 1,
               maxBitrateKbps: 600,
               attendeeId: 'xy1',
+              externalUserId: 'ext#xy1',
               mediaType: SdkStreamMediaType.AUDIO,
             }),
             new SdkStreamDescriptor({
@@ -247,6 +253,7 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 399,
               maxBitrateKbps: 800,
               attendeeId: 'xy1',
+              externalUserId: 'ext#xy1',
               mediaType: SdkStreamMediaType.VIDEO,
             }),
             new SdkStreamDescriptor({
@@ -254,6 +261,7 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 2,
               maxBitrateKbps: 50,
               attendeeId: 'xy2',
+              externalUserId: 'ext#xy2',
               mediaType: SdkStreamMediaType.VIDEO,
             }),
             new SdkStreamDescriptor({
@@ -261,6 +269,7 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 399,
               maxBitrateKbps: 200,
               attendeeId: 'xy1',
+              externalUserId: 'ext#xy1',
               mediaType: SdkStreamMediaType.VIDEO,
             }),
             new SdkStreamDescriptor({
@@ -268,16 +277,20 @@ describe('DefaultVideoStreamIndex', () => {
               groupId: 2,
               maxBitrateKbps: 400,
               attendeeId: 'xy2',
+              externalUserId: 'ext#xy2',
               mediaType: SdkStreamMediaType.VIDEO,
             }),
           ],
         })
       );
-      expect(Array.from(index.allVideoSendingAttendeesExcludingSelf('xy1')).sort()).to.deep.equal(
-        ['xy2'].sort()
+      expect(index.allVideoSendingAttendeesExcludingSelf('xy1').sort()).to.deep.equal(
+        [{ attendeeId: 'xy2', externalUserId: 'ext#xy2' }].sort(compare)
       );
       expect(Array.from(index.allVideoSendingAttendeesExcludingSelf('xy3')).sort()).to.deep.equal(
-        ['xy1', 'xy2'].sort()
+        [
+          { attendeeId: 'xy1', externalUserId: 'ext#xy1' },
+          { attendeeId: 'xy2', externalUserId: 'ext#xy2' },
+        ].sort(compare)
       );
     });
   });