+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make AWS SDK for Java camelCased meeting-attendee response compatible with Chime SDK for JavaScript
 - Mark InvalidSequenceTokenExceptions as warning
 - Add an optional parameter to the serverless demo deployment script to specify Chime endpoint, and deploy to a new devo stage that talks to gamma Chime endpoint for canary
+- Add simulcast uplink policy layer change notification methods and observers
+- Add remote videos available observer and method
 
 ### Changed
 - Update test results to Sauce Labs before emitting CloudWatch metrics for integration tests